+++ src/index.ts
@@ -19,7 +19,7 @@ export { json, jsonArray, jsonObject } from './json';
 export { instanceOf, lazy, prep } from './misc';
 export { anyNumber, integer, number, positiveInteger, positiveNumber } from './numbers';
 export { bigint } from './numbers';
-export { dict, exact, inexact, mapping, object, pojo } from './objects';
+export { dict, exact, inexact, mapping, object, pojo, record } from './objects';
 export { nonEmptyString, regex, string } from './strings';
 export { email, httpsUrl, url, uuid, uuidv1, uuidv4 } from './strings';
 export { decimal, hexadecimal, numeric } from './strings';