+++ test-d/inference.test-d.ts
@@ -46,6 +46,7 @@ import {
   positiveInteger,
   positiveNumber,
   prep,
+  record,
   regex,
   select,
   set,
@@ -318,6 +319,7 @@ expectType<unknown>(test(inexact({})).b);
 
 expectType<Record<string, unknown>>(test(pojo));
 expectType<Map<string, number>>(test(mapping(number)));
+expectType<Record<string, number>>(test(record(number)));
 expectType<Record<string, number>>(test(dict(number)));
 
 expectType<string>(test(lazy(() => string)));